enviro_digital_twin/lambdas/aurora_processing_lambda/pyproject.toml
[tool.poetry]
name = "aurora_processing_lambda"
version = "0.1.0"
description = ""
authors = ["powers_rc <powerscameron@ymail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.26.21"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
